+++ src/routes.rs
@@ -1,12 +1,8 @@
 use crate::models::{
     ApiResponse, ErrorResponse, SolanaProgramBuild, SolanaProgramBuildParams, Status,
     SuccessResponse, VerificationStatusParams, VerificationStatusResponse, VerifyAsyncResponse,
-    VerifySyncResponse,
 };
-use crate::operations::{
-    check_is_build_params_exists_already, check_is_program_verified_within_24hrs, verify_build,
-};
-use crate::state::AppState;
+use crate::db::DbClient;
 use axum::extract::Path;
 use axum::{
     extract::State,
@@ -16,13 +12,12 @@ use axum::{
 use chrono::Utc;
 use serde_json::{json, Value};
 
-pub fn create_router(app_state: AppState) -> Router {
+pub fn create_router(db: DbClient) -> Router {
     Router::new()
         .route("/", get(index))
         .route("/verify", post(handle_verify))
-        .route("/verify_sync", post(handle_verify_sync))
         .route("/status/:address", get(handle_verify_status))
-        .with_state(app_state)
+        .with_state(db)
 }
 
 async fn index() -> Json<Value> {
@@ -46,7 +41,7 @@ async fn index() -> Json<Value> {
 
 // Route handler for POST /verify which creates a new process to verify the program
 async fn handle_verify(
-    State(app): State<AppState>,
+    State(db): State<DbClient>,
     Json(payload): Json<SolanaProgramBuildParams>,
 ) -> Json<ApiResponse> {
     let verify_build_data = SolanaProgramBuild {
@@ -60,7 +55,7 @@ async fn handle_verify(
     };
 
     // First check if the program is already verified
-    let is_exists = check_is_build_params_exists_already(app.clone(), &payload).await;
+    let is_exists = db.check_is_build_params_exists_already(&payload).await;
 
     if let Ok(is_exists) = is_exists {
         if is_exists {
@@ -72,17 +67,14 @@ async fn handle_verify(
     }
 
     // insert into database
-    let insert = app
-        .db_client
-        .insert_or_update_build(&verify_build_data)
-        .await;
+    let insert = db.insert_or_update_build(&verify_build_data).await;
 
     match insert {
         Ok(_) => {
             tracing::info!("Inserted into database");
             //run task in background
             tokio::spawn(async move {
-                let _ = verify_build(app, payload).await;
+                let _ = db.verify_build(payload).await;
             });
 
             Json(ApiResponse::Success(SuccessResponse::VerifyAsync(
@@ -103,97 +95,94 @@ async fn handle_verify(
 }
 
 // Route handler for POST /sync_verify which creates a new process to verify the program synchronously
-async fn handle_verify_sync(
-    State(app): State<AppState>,
-    Json(payload): Json<SolanaProgramBuildParams>,
-) -> Json<ApiResponse> {
-    let verify_build_data = SolanaProgramBuild {
-        id: uuid::Uuid::new_v4().to_string(),
-        repository: payload.repository.clone(),
-        commit_hash: payload.commit_hash.clone(),
-        program_id: payload.program_id.clone(),
-        lib_name: payload.lib_name.clone(),
-        bpf_flag: payload.bpf_flag.unwrap_or(false),
-        created_at: Utc::now().naive_utc(),
-    };
+// async fn handle_verify_sync(
+//     State(db): State<DbClient>,
+//     Json(payload): Json<SolanaProgramBuildParams>,
+// ) -> Json<ApiResponse> {
+//     let verify_build_data = SolanaProgramBuild {
+//         id: uuid::Uuid::new_v4().to_string(),
+//         repository: payload.repository.clone(),
+//         commit_hash: payload.commit_hash.clone(),
+//         program_id: payload.program_id.clone(),
+//         lib_name: payload.lib_name.clone(),
+//         bpf_flag: payload.bpf_flag.unwrap_or(false),
+//         created_at: Utc::now().naive_utc(),
+//     };
 
-    // First check if the program is already verified
-    let is_exists = check_is_build_params_exists_already(app.clone(), &payload).await;
+//     // First check if the program is already verified
+//     let is_exists = check_is_build_params_exists_already(&db, &payload).await;
 
-    if let Ok(is_exists) = is_exists {
-        if is_exists {
-            return Json(ApiResponse::Error(ErrorResponse {
-                status: Status::Error,
-                error: "We have already processed this request".to_string(),
-            }));
-        }
-    }
+//     if let Ok(is_exists) = is_exists {
+//         if is_exists {
+//             return Json(ApiResponse::Error(ErrorResponse {
+//                 status: Status::Error,
+//                 error: "We have already processed this request".to_string(),
+//             }));
+//         }
+//     }
 
-    // Else insert into database
-    let insert = app
-        .db_client
-        .insert_or_update_build(&verify_build_data)
-        .await;
+//     // Else insert into database
+//     let insert = db.insert_or_update_build(&verify_build_data).await;
 
-    match insert {
-        Ok(_) => {
-            tracing::info!("Inserted into database");
-            // Run task in background
-            let handle = tokio::task::spawn_blocking(move || verify_build(app, payload));
+//     match insert {
+//         Ok(_) => {
+//             tracing::info!("Inserted into database");
+//             // Run task in background
+//             let handle = tokio::task::spawn_blocking(move || verify_build(&db, payload));
 
-            let task = handle.await;
+//             let task = handle.await;
 
-            if let Ok(res) = task {
-                match res.await {
-                    Ok(verified_program) => {
-                        tracing::info!("Build verification completed");
-                        Json(ApiResponse::Success(SuccessResponse::VerifySync(
-                            VerifySyncResponse {
-                                status: Status::Success,
-                                is_verified: verified_program.is_verified,
-                                on_chain_hash: verified_program.on_chain_hash,
-                                executable_hash: verified_program.executable_hash,
-                                message: if verified_program.is_verified {
-                                    "On chain program verified".to_string()
-                                } else {
-                                    "On chain program not verified".to_string()
-                                },
-                            },
-                        )))
-                    }
-                    Err(e) => {
-                        tracing::error!("Error verifying build: {:?}", e);
-                        Json(ApiResponse::Error(ErrorResponse {
-                            status: Status::Error,
-                            error: format!("unexpected error occurred {:?}", e),
-                        }))
-                    }
-                }
-            } else {
-                let err_info = format!("unexpected error occurred {:?}", task.err());
-                tracing::error!(err_info);
-                Json(ApiResponse::Error(ErrorResponse {
-                    status: Status::Error,
-                    error: err_info,
-                }))
-            }
-        }
-        Err(e) => {
-            tracing::error!("Error inserting into database: {:?}", e);
-            Json(ApiResponse::Error(ErrorResponse {
-                status: Status::Error,
-                error: "unexpected error occurred".to_string(),
-            }))
-        }
-    }
-}
+//             if let Ok(res) = task {
+//                 match res.await {
+//                     Ok(verified_program) => {
+//                         tracing::info!("Build verification completed");
+//                         Json(ApiResponse::Success(SuccessResponse::VerifySync(
+//                             VerifySyncResponse {
+//                                 status: Status::Success,
+//                                 is_verified: verified_program.is_verified,
+//                                 on_chain_hash: verified_program.on_chain_hash,
+//                                 executable_hash: verified_program.executable_hash,
+//                                 message: if verified_program.is_verified {
+//                                     "On chain program verified".to_string()
+//                                 } else {
+//                                     "On chain program not verified".to_string()
+//                                 },
+//                             },
+//                         )))
+//                     }
+//                     Err(e) => {
+//                         tracing::error!("Error verifying build: {:?}", e);
+//                         Json(ApiResponse::Error(ErrorResponse {
+//                             status: Status::Error,
+//                             error: format!("unexpected error occurred {:?}", e),
+//                         }))
+//                     }
+//                 }
+//             } else {
+//                 let err_info = format!("unexpected error occurred {:?}", task.err());
+//                 tracing::error!(err_info);
+//                 Json(ApiResponse::Error(ErrorResponse {
+//                     status: Status::Error,
+//                     error: err_info,
+//                 }))
+//             }
+//         }
+//         Err(e) => {
+//             tracing::error!("Error inserting into database: {:?}", e);
+//             Json(ApiResponse::Error(ErrorResponse {
+//                 status: Status::Error,
+//                 error: "unexpected error occurred".to_string(),
+//             }))
+//         }
+//     }
+// }
 
 //  Route handler for GET /status/:address which checks if the program is verified or not
 async fn handle_verify_status(
-    State(app): State<AppState>,
+    State(db): State<DbClient>,
     Path(VerificationStatusParams { address }): Path<VerificationStatusParams>,
 ) -> Json<ApiResponse> {
-    let result = check_is_program_verified_within_24hrs(app, address).await;
+    let result = db.check_is_program_verified_within_24hrs(address).await;
 
     if let Ok(result) = result {
         return Json(ApiResponse::Success(SuccessResponse::VerificationStatus(
